Fixed email validator patterns
Also moved form specs into the correct projects.
